fix(security): corrige 6 findings code review (sec + arch)
- utils/date.ts: ajoute isWithinBusinessHours (extrait de index.ts)
- appointments/index.ts: importe depuis utils/date (supprime doublons)
- appointments/[id].ts: valide UUID path param, isWithinBusinessHours sur reschedule,
  expire ancien Payment Link Stripe avant report, importe isWithinBusinessHours
- contact.ts + appointments/index.ts: IP via x-nf-client-connection-ip (anti-spoofing)
- rate-limit.ts: documente limites cold-start / multi-instance Netlify
- mes-rdvs.astro: select colonnes explicites (RGPD, évite sérialisation PII)
- 001_init.sql: documente UNIQUE NULL behavior sur stripe_event_id
